webServerHandler now displays content while app serves as wrapper
